[LUA]Move PUP JA to jobUtils file
[LUA]Move PUP JA to jobUtils file

[LUA]Move PUP JA to jobUtils file

[LUA]Move PUP JA to jobUtils file
