+++ scripts/actions/abilities/activate.lua
@@ -9,27 +9,11 @@
 local abilityObject = {}
 
 abilityObject.onAbilityCheck = function(player, target, ability)
-    if player:getPet() ~= nil then
-        return xi.msg.basic.ALREADY_HAS_A_PET, 0
-    elseif not player:canUseMisc(xi.zoneMisc.PET) then
-        return xi.msg.basic.CANT_BE_USED_IN_AREA, 0
-    elseif player:isExceedingElementalCapacity() then
-        return xi.msg.basic.AUTO_EXCEEDS_CAPACITY, 0
-    end
-
-    return 0, 0
+    return xi.job_utils.puppetmaster.onAbilityCheckActivate(player, target, ability)
 end
 
 abilityObject.onUseAbility = function(player, target, ability)
-    xi.pet.spawnPet(player, xi.petId.AUTOMATON)
-
-    local pet = player:getPet()
-
-    if pet then
-        local jpValue = player:getJobPointLevel(xi.jp.AUTOMATON_HP_MP_BONUS)
-        pet:addMod(xi.mod.HP, jpValue * 10)
-        pet:addMod(xi.mod.MP, jpValue * 5)
-    end
+    return xi.job_utils.puppetmaster.onAbilityUseActivate(player, target, ability)
 end
 
 return abilityObject
