+++ scripts/actions/abilities/deus_ex_automata.lua
@@ -9,28 +9,11 @@
 local abilityObject = {}
 
 abilityObject.onAbilityCheck = function(player, target, ability)
-    if player:getPet() ~= nil then
-        return xi.msg.basic.ALREADY_HAS_A_PET, 0
-    elseif not player:canUseMisc(xi.zoneMisc.PET) then
-        return xi.msg.basic.CANT_BE_USED_IN_AREA, 0
-    else
-        local jpValue = player:getJobPointLevel(xi.jp.DEUS_EX_AUTOMATA_RECAST)
-
-        ability:setRecast(ability:getRecast() - jpValue)
-
-        return 0, 0
-    end
+    return xi.job_utils.puppetmaster.onAbilityCheckDeuxExAutomata(player, target, ability)
 end
 
 abilityObject.onUseAbility = function(player, target, ability)
-    xi.pet.spawnPet(player, xi.petId.AUTOMATON)
-    local pet = player:getPet()
-
-    if pet then
-        local percent = math.floor((player:getMainLvl() / 3)) / 100
-        pet:setHP(math.max(pet:getHP() * percent, 1))
-        pet:setMP(pet:getMP() * percent)
-    end
+    return xi.job_utils.puppetmaster.onAbilityUseDeuxExAutomata(player, target, ability)
 end
 
 return abilityObject
