+++ scripts/actions/abilities/retrieve.lua
@@ -9,11 +9,11 @@
 local abilityObject = {}
 
 abilityObject.onAbilityCheck = function(player, target, ability)
-    return 0, 0
+    return xi.job_utils.puppetmaster.onAbilityCheckRetrieve(player, target, ability)
 end
 
 abilityObject.onUseAbility = function(player, target, ability)
-    player:petRetreat()
+    return xi.job_utils.puppetmaster.onAbilityUseRetrieve(player, target, ability)
 end
 
 return abilityObject
