+++ scripts/actions/abilities/role_reversal.lua
@@ -9,27 +9,11 @@
 local abilityObject = {}
 
 abilityObject.onAbilityCheck = function(player, target, ability)
-    local pet = player:getPet()
-
-    if not pet then
-        return xi.msg.basic.REQUIRES_A_PET, 0
-    elseif not pet:isAutomaton() then
-        return xi.msg.basic.NO_EFFECT_ON_PET, 0
-    else
-        return 0, 0
-    end
+    return xi.job_utils.puppetmaster.onAbilityCheckRoleReversal(player, target, ability)
 end
 
 abilityObject.onUseAbility = function(player, target, ability)
-    local pet = player:getPet()
-    if pet then
-        local bonus    = 1 + (player:getMerit(xi.merit.ROLE_REVERSAL) - 5) / 100
-        local playerHP = player:getHP()
-        local petHP    = pet:getHP()
-
-        pet:setHP(math.max(playerHP * bonus, 1))
-        player:setHP(math.max(petHP * bonus, 1))
-    end
+    return xi.job_utils.puppetmaster.onAbilityUseRoleReversal(player, target, ability)
 end
 
 return abilityObject
